refactor: ladder 도메인 패키지 생성
