+++ src/main/java/ladder/domain/ladder/DirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 public interface DirectionGenerator {
     Direction generateFirst();
