+++ src/test/java/ladder/domain/ladder/LadderLineTest.java
@@ -1,11 +1,11 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static ladder.domain.LadderLine.PERSON_COUNT_MUST_MORE_THEN_TWO;
+import static ladder.domain.ladder.LadderLine.PERSON_COUNT_MUST_MORE_THEN_TWO;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
